+++ src/index.ts
@@ -1,7 +1,4 @@
 import {DFA} from "./automata/regular/DFA";
+import {NFA} from "./automata/regular/NFA";
 
-export type char = string & { length: 1 }
-export const toChar = (str:string) => (str[0] as char);
-export const EPSILON:char = toChar('ε');
-
-export {DFA}
+export {DFA, NFA}
